Fixes Issue#1740. More fixes to pytests
